backend/requirements.txt
# requirements.txt - БЕЗ ПРОБЛЕМ С RUST
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.13.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.25.1

# Используйте эту версию Pydantic
pydantic==1.10.13

python-dotenv==1.0.0